requirements.txt
requests
pycryptodome
tabulate